Split generic behavior association tests
